+++ codes.py
@@ -1,0 +1,8 @@
+ok = 0
+error = 1
+model_validation_error = 2
+
+codenames: dict[int, str] = {
+    1: "error",
+    2: "model_validation_error",
+}